==> 01 - Scripts/01 - Filtering Dataset to Scope of Project.ipynb <==
# TOC
1. Import Libraries & Data     
2. Inspecting original data   
3. Converting date column to datetime format   
4. Filtering data to scope of project (Jan 1 2018 through December 31 2022)   
5. Saving filtered data to pkl   



# Importing libraries
import pandas as pd
import numpy as np
import os
import datetime

# Importing original data
path = r'C:\Users\Ryzen RGB Madness!!!\Iowa Liquor Sales Analysis'

orig_data = pd.read_csv(os.path.join(path, '01 - Data', 'Original', 'Iowa_Liquor_Sales_20231118.csv'))

# Investigating original data
orig_data.shape

orig_data.dtypes

# Converting date column to datetime type
orig_data['date'] = pd.to_datetime(orig_data['date'], format="%m/%d/%Y")

orig_data.dtypes

orig_data.head()

orig_data.shape

filtered_data = orig_data[(orig_data['date'] >= '2018-01-01') & (orig_data['date'] <= '2022-12-31')]

filtered_data.shape

# Saving filtered data prior to cleaning (the original dataset crashes my computer)
filtered_data.to_pickle(os.path.join(path, '01 - Data', 'Original', 'iowa_liquor_sales_jan18_to_dec22_orig.pkl'))
