==> 01 - Scripts/02 - Cleaning & Wrangling Filtered Data.ipynb <==
# TOC
1. Import Libraries & Data     
2. Inspecting filtered data   
3. Renaming columns to best practices (no capitalization, special characters)   
4. Converting date column to datetime format   
5. Checking for null values   
6. Separating 'zip_code' nulls into csv for updating   
7. Separating 'category' nulls into csv for updating  
8. Dropping 'store_location' and 'county_num' columns (not needed)
9. Dropping 'zip_code' nulls & 'category' nulls from main data (merge prep)  
10. Setting correct datatypes   
11. Merging dataframes via pd.concat & checking null values     
12. Resetting dataframe index   
13. Correcting minor issues (removing commas from address fields, correcting misspelled city names)   
14. Exporting final cleaned dataframe to pickle to begin analysis. 

# Importing libraries
import pandas as pd
import numpy as np
import os
import datetime

# Importing filtered data
path = r'C:\Users\Ryzen RGB Madness!!!\Iowa Liquor Sales Analysis'

liquor_data = pd.read_pickle(os.path.join(path, '01 - Data', 'Original', 'iowa_liquor_sales_jan18_to_dec22_orig.pkl'))

# Inspecting filtered data
liquor_data.shape

# Checking for null values
liquor_data.isnull().sum()

# Separating zip code nulls into separate dataframe
address_issues = liquor_data[liquor_data['zip_code'].isnull()==True]

address_issues.shape

# Saving missing addresses to csv to impute data (check against records in main dataset)
address_issues.to_csv(os.path.join(path, '01 - Data', 'Original', 'iowa_liquor_sales_jan18_to_dec22_missing_addys2.csv'))

# Separating category nulls into separate dataframe
category_issues = liquor_data[liquor_data['category'].isnull()==True]

category_issues.shape

# Saving missing categories to csv to impute data (check against records in main dataset)
category_issues.to_csv(os.path.join(path, '01 - Data', 'Original', 'iowa_liquor_sales_jan18_to_dec22_missing_cats2.csv'))

# Removing nulls
liquor_data_updated = liquor_data.dropna(axis=0,subset=['address'])

liquor_data_updated.shape

# Checking for null values
liquor_data_updated.isnull().sum()

liquor_data_updated = liquor_data.dropna(axis=0,subset=['zip_code'])

liquor_data_updated.shape

# Checking for null values
liquor_data_updated.isnull().sum()

liquor_data_updated = liquor_data_updated.dropna(axis=0,subset=['category'])

liquor_data_updated.shape

# Checking for null values
liquor_data_updated.isnull().sum()

# Removing store_location & county_num from main dataframe
liquor_data_updated = liquor_data_updated.drop(columns=['store_location','county_num'])

# Importing updated csv data
updated_info = pd.read_csv(os.path.join(path, '01 - Data', 'Original', 'iowa_liquor_sales_updated_info.csv'))

updated_info.shape

liquor_data_updated.shape

liquor_data_updated.describe()

updated_info.dtypes

# Converting updated_info 'date' column to datetime
updated_info['date'] = pd.to_datetime(updated_info['date'], format="%m/%d/%Y")

updated_info.dtypes

updated_info['date'].describe()

# Removing index column
updated_info.dtypes

updated_info = updated_info.drop('Unnamed: 0', axis=1)

updated_info.columns

# Merging dataframes (test)
pd.concat([liquor_data_updated, updated_info], axis=0)

# Merging dataframes into new dataframe
liquor_data_complete = pd.concat([liquor_data_updated, updated_info], axis=0)

liquor_data_complete.shape

liquor_data_complete['date'].describe()

# Resetting index
liquor_data_complete.reset_index(inplace=True, drop=True)

liquor_data_complete.dtypes

liquor_data_complete.describe()

# Correcting data types
convert_dict = {'invoice_number': str,
                'store_number': str,
                'store_name': str,
                'address': str,
                'city': str,
                'zip_code': str,
                'county': str,
                'category': str,
                'category_name': str,
                'vendor_number': str,
                'vendor_name': str,
                'item_number': str,
                'item_description': str                
                }

liquor_data_complete = liquor_data_complete.astype(convert_dict)

liquor_data_complete.dtypes

# Converting non-numerical number identifiers (store number, category, vendor number, item number) to string
liquor_data_complete['store_number'] = liquor_data_complete['store_number'].astype('string')

liquor_data_complete.dtypes

liquor_data_complete['category'] = liquor_data_complete['category'].astype('string')

liquor_data_complete['vendor_number'] = liquor_data_complete['vendor_number'].astype('string')

liquor_data_complete['item_number'] = liquor_data_complete['item_number'].astype('string')

liquor_data_complete.dtypes

# Removing commas from address column
liquor_data_complete['address'] = liquor_data_complete['address'].str.replace(","," ")

# Checking commas
liquor_data_complete.loc[liquor_data_complete['invoice_number'] == 'INV-14279100033']

# Correcting misspelled city
liquor_data_complete['city'] = liquor_data_complete['city'].str.replace("OTUMWA","OTTUMWA")

# Checking city error
liquor_data_complete.loc[liquor_data_complete['city'] == 'OTUMWA']

# Correcting misspelled city
liquor_data_complete['city'] = liquor_data_complete['city'].str.replace("ARNOLD'S PARK","ARNOLDS PARK")

# Checking city error
liquor_data_complete.loc[liquor_data_complete['city'] == "ARNOLD'S PARK"]

# Verifying dates
liquor_data_complete['date'].describe()

# Exporting complete cleaned dataframe
liquor_data_complete.to_pickle(os.path.join(path, '01 - Data', 'Cleaned', 'iowa_liquor_data_complete.pkl'))
