==> 01 - Scripts/03 - Descriptive Statistics.ipynb <==
# TOC
1. Importing libraries & data   
2. Descriptive statistics - quantitative   
3. Descriptive statistics - categorical

# Importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn
import scipy

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Iowa Liquor Sales Analysis'

liquor_data_complete = pd.read_pickle(os.path.join(path, '01 - Data', 'Cleaned', 'iowa_liquor_data_complete.pkl'))

liquor_data_complete.dtypes

# Descriptive statistics on all numeric fields
liquor_data_complete['pack'].describe()

liquor_data_complete['bottle_volume_ml'].describe()

liquor_data_complete['state_bottle_cost'].describe()

liquor_data_complete['state_bottle_retail'].describe()

# What is that $18,000 bottle of liquor??
cost_max = liquor_data_complete[liquor_data_complete['state_bottle_cost']==18436.0]

cost_max.head()

liquor_data_complete['bottles_sold'].describe()

# What is that bottles_sold max??
bottle_max = liquor_data_complete[liquor_data_complete['bottles_sold']==13200.0]

bottle_max.head()

liquor_data_complete['sale_dollars'].describe()

# What is that sale dollars max??? 
sale_max = liquor_data_complete[liquor_data_complete['sale_dollars']==279557.28]

sale_max.head()

liquor_data_complete['volume_sold_liters'].describe()

liquor_data_complete['volume_sold_gallons'].describe()

liquor_data_complete['store_number'].describe()

liquor_data_complete['store_name'].describe()

liquor_data_complete['address'].describe()

liquor_data_complete['city'].describe()

liquor_data_complete['zip_code'].describe()

liquor_data_complete['county'].describe()

liquor_data_complete['category'].describe()

liquor_data_complete['category_name'].describe()

liquor_data_complete['vendor_number'].describe()

liquor_data_complete['vendor_name'].describe()

liquor_data_complete['item_number'].describe()

liquor_data_complete['item_description'].describe()

liquor_data_complete['date'].describe()

liquor_data_complete['invoice_number'].describe()
