==> 01 - Scripts/05 - Updating Location Information.ipynb <==
# TOC
1. Importing library & main data   
2. Matching county names to JSON format   
3. Importing county FIPS data   
4. Merging FIPS data into main dataframe   
5. Separating address information into new dataframe   
6. Exporting unique addresses   
7. Importing store latitude/longitude data   
8. Merging latitude/longitude data into main dataframe   

# Importing libraries    
import pandas as pd
import numpy as np
import os

# Importing data
path = r'C:\Users\Ryzen RGB Madness!!!\Iowa Liquor Sales Analysis'

liquor_data_complete = pd.read_pickle(os.path.join(path, '01 - Data', 'Cleaned', 'iowa_liquor_sales_and_population.pkl'))

# Setting county names into proper capitalization to match JSON data
liquor_data_complete['county'] = liquor_data_complete['county'].str.title()

# Matching O'Brien County to FIPS & JSON data
liquor_data_complete['county'] = liquor_data_complete['county'].str.replace("O'Brien","Obrien")

liquor_data_complete.head()

# Importing FIPS data
fips_data = pd.read_csv(os.path.join(path, '01 - Data', 'Original', 'iowa_county_fips.csv'))

fips_data.head()

# Merging FIPS data into main dataframe
liquor_data_complete = liquor_data_complete.merge(fips_data[['county_fips','county']],left_on=['county'],right_on=['county'],how='left')

liquor_data_complete.columns

# Nulls?
liquor_data_complete.isnull().sum()

fips_nulls = liquor_data_complete[liquor_data_complete['county_fips'].isnull()==True]

fips_nulls.head()

# Moving fips column before county 
liquor_data_complete = liquor_data_complete[['invoice_number', 'date', 'store_number', 'store_name', 'address',
       'city', 'zip_code', 'county_fips', 'county', 'population', 'category', 'category_name',
       'vendor_number', 'vendor_name', 'item_number', 'item_description',
       'pack', 'bottle_volume_ml', 'state_bottle_cost', 'state_bottle_retail',
       'bottles_sold', 'sale_dollars', 'volume_sold_liters',
       'volume_sold_gallons']]

# Duplicating dataframe (lat/long prep)
addresses = liquor_data_complete[['address', 'city', 'zip_code']].copy()

addresses.shape

# Unique addresses only
unique_addys = pd.DataFrame(addresses.address.dropna().unique(), columns=['address'])

unique_addys.shape

# exporting unique addresses to csv
unique_addys.to_csv(os.path.join(path, '01 - Data', 'Cleaned', 'iowa_store_lats_longs.csv'))

#### Outside Python:
1. Verified no duplicates (Excel)   
2. Added state column (Excel)   
3. Cleaned unnecessary information from address column, notated row and what was (temporarily) removed (Excel)   
4. Uploaded spreadsheet to Geocodio to get latitude/longitude data (Geocodio)   
5. Saved original Geocodio csv with all data (including accuracy and source)   
6. Removed unnecessary columns that don't need to be appended into dataset (Excel)   
7. Saved revised csv as iowa_data_lats_longs

# Importing lat/long data
latlong_data = pd.read_csv(os.path.join(path, '01 - Data', 'Cleaned', 'iowa_data_lats_longs.csv'))

latlong_data.shape

latlong_data.head()

# Merging lat/long data into main dataframe
liquor_data_all = liquor_data_complete.merge(latlong_data[['address', 'latitude', 'longitude']],left_on=['address'],right_on=['address'],how='left')

liquor_data_all.shape

liquor_data_all.head()

# Moving lat/long columns before county_fips
liquor_data_all = liquor_data_all[['invoice_number', 'date', 'store_number', 'store_name', 'address',
       'city', 'zip_code', 'latitude', 'longitude', 'county_fips', 'county', 'population', 'category', 'category_name',
       'vendor_number', 'vendor_name', 'item_number', 'item_description',
       'pack', 'bottle_volume_ml', 'state_bottle_cost', 'state_bottle_retail',
       'bottles_sold', 'sale_dollars', 'volume_sold_liters',
       'volume_sold_gallons']]

# Checking for nulls
liquor_data_all.isnull().sum()

# Exporting full dataset
liquor_data_all.to_pickle(os.path.join(path, '01 - Data', 'Cleaned', 'iowa_sales_location_population.pkl'))
